Fix stack overflow in llvm::Verifier::visitMDNode()
This code is a proof-of-concept more than anything else. It is not ready
to be merged in its current state. Any version of this fix must somehow
measure the depth of the metadata nodes (or a rough upper bound), and it
must somehow send that depth to the LLVM codegen entry point. There is
likely a tradeoff to be made between performance and readability here.