Fix stack overflow in rustc_ast::ptr::P<_> destructor
In rustc_ast, an Expr contains an ExprKind, which can contain a P<Expr>.
To prevent the stack from overflowing when dropping an Expr, one of the
types must use ManuallyDrop for its fields. Since both Expr and ExprKind
are frequently deconstructed into their fields using match expressions,
I chose to attach it to P::ptr.