Allow uninhabited types in generator interiors
Drop tracking in the generator interior type checking pass would count
all values in unreachable code as dropped (e.g. code after a call to a
function with an uninhabited return type), which would lead to those
values not being included in the witness type. This resulted in the type
checker being more precise than the corresponding sanity check in the
MIR transform.

This patch changes the check in the MIR code to match the results of
typeck by skipping uninhabited types.