Allow source link to wrap onto next line
Like the out-of-band information at the top of the page, this uses
flex-wrap: wrap to allow `source` to move to the next line when space is
tight. Also, on mobile layouts move it to the next line always, for a
more consistent interface.

This allows more of the method signature to be displayed on a single
line, and avoids confusing `source` with one of the items in the method
signature.

Also tweak the padding on summary items slightly so they line up
properly with their docblock.