Change `AsFd::as_fd` etc. from `&self` to `self`.
Change the `AsFd` trait from:

```rust
trait AsFd {
    fn as_fd(&self) -> BorrowedFd<'_>;
}
```

to:

```rust
trait AsFd<'a> {
    fn as_fd(self) -> BorrowedFd<'a>;
}
```

and similar for `AsHandle` and `AsSocket`. Instead of implementing `AsFd`
for `OwnedFd` etc., it's now implemented for `&'a OwnedFd` etc.

This means that implementing `AsFd` changes from:

```rust
impl AsFd for MyType {
   fn as_fd(&self) -> BorrowedFd<'_> {
      ...
   }
}
```

to:

```rust
impl<'a> AsFd<'a> for &'a MyType {
   fn as_fd(self) -> BorrowedFd<'a> {
      ...
   }
}
```

Defining a function that uses an `AsFd` changes from:

```rust
fn frob<Fd: AsFd>(fd: &Fd)
```

to:

```rust
fn frob<'a, Fd: AsFd<'a>>(fd: Fd)
```

It also means that users of such functions can pass a `BorrowedFd` to them
directly, instead of having to write an extra `&` to conceptually add an
extra borrow around `BorrowedFd`.

In total, it means somewhat more typing for implementors of the traits and
implementors of functions that take `AsFd` parameters. But it also
avoids a common confusion, and allows users to have fewer `&`s. And, it
means fewer "reference to reference"-like situations.

As an example of this in practice, the new unstable `fchown` function in std
is [already taking an `AsFd` by value instead of by reference]. As
written, it can't be passed a `&File`, and passing it a `File` consumes
and closes the file. With this PR, it accepts `&File` and never consumes
anything.

One subtle detail is that the Unix internal `File`, `Socket`, and
`AnonPipe` types need to stop implementing `AsFd`, so that they don't
require stability attributes which are otherwise apparently required.
Since these are internal types, this doesn't affect the public API.
There are comments in the code for these.

Another subtle detail is that with the convention of types implementing
`AsFd` for `&'a T`, they don't implement it for `&'a mut T`. For example,
with the `frob` function above, `frob(&mut file)` gets this error:

```
error[E0277]: the trait bound `&mut File: AsFd<'_>` is not satisfied
```

One needs to use `frob(&file)` instead. If one has a `mut T`, it's just
a matter of typing `&` instead of `&mut`, but if one has a `&mut T`, one
needs to use `&*`. This is a little awkward, however the only alternative
I know about is to oblige all types that implement `AsFd<'a>` for `&'a T`
to also implement `AsFd<'a>` for `&'a mut T`, which seems worse.

[already taking an `AsFd` by value instead of by reference]: https://github.com/rust-lang/rust/blob/734368a200904ef9c21db86c595dc04263c87be0/library/std/src/os/unix/fs.rs#L974