Fix stack protector basic test
This is a >= condition, so we need a maximum size of 7 to not
create a stack protector in basic mode.

The reason this still worked is that the alloca type was converted
into an integer (rather than an array). The way these heuristics
are implemented in LLVM is rather questionable and not resilient
to optimization.