Drop incremental compilation support
This removes the majority (all that I could easily find) of incremental support
from the compiler. This is not intended to land on master - but is rather used
for perf to evaluate the impact incremental has on non-incremental builds, which
may provide useful insight into where we can better optimize the conditionals
and such that gate incremental today.